requirements.txt
alembic==1.4.2
altgraph==0.17
aniso8601==8.0.0
astroid==2.4.2
bcrypt==3.1.7
beautifulsoup4==4.8.1
bokeh==1.4.0
bs4==0.0.1
cachelib==0.1.1
cachetools==4.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9.2
dash==1.11.0
dash-core-components==1.9.1
dash-html-components==1.0.3
dash-renderer==1.4.0
dash-table==4.6.2
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.0
fanstatic==1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Compress==1.4.0
Flask-Dance==3.0.0
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-RESTful==0.3.8
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.18.2
google-api-core==1.17.0
google-api-python-client==1.8.0
google-auth==1.14.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.2
idna==2.8
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.16.0
lazy-object-proxy==1.4.3
lxml==4.5.0
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
Naked==0.1.31
nltk==3.5
numpy==1.19.0
oauth2client==4.1.3
oauthlib==3.1.0
packaging==19.2
pandas==0.25.3
pdfminer.six==20200402
pefile==2019.4.18
Pillow==7.1.2
plotly==4.6.0
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.7
PyDrive==1.3.1
pygame==1.9.6
pyjsparser==2.7.1
PyJWT==1.4.2
pylint==2.6.0
pyparsing==2.4.6
PyPDF2==1.26.0
pypiwin32==223
PyQt5==5.14.2
PyQt5-sip==12.7.2
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==5.2
redis==3.5.3
regex==2020.6.8
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
shellescape==3.8.1
shutilwhich==1.1.0
six==1.13.0
sortedcontainers==2.1.0
soupsieve==1.9.5
SQLAlchemy==1.3.16
SQLAlchemy-Utils==0.36.6
tornado==6.0.3
torpy==1.0.10
tqdm==4.47.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.7
URLObject==2.4.3
visitor==0.1.3
WebOb==1.8.6
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.1
